Fixed a bug with setting the time of sending messages> Replaced Router with HashRouter. Distributed the files into folders.
