+++ med-web-app-front/src/components/messenger/recipient.msg.component.js
