+++ med-web-app-front/src/components/messenger/sender.msg.component.js
