--- med-web-app-front/src/components/questions.component.js
@@ -1,43 +1,0 @@
-import React, { Component } from "react";
-
-import TestService from "../services/test.service";
-
-export default class BoardUser extends Component {
-    constructor(props) {
-        super(props);
-
-        this.state = {
-            content: ""
-        };
-    }
-
-    componentDidMount() {
-        TestService.getUserBoardContent().then(
-            response => {
-                this.setState({
-                    content: response.data
-                });
-            },
-            error => {
-                this.setState({
-                    content:
-                        (error.response &&
-                            error.response.data &&
-                            error.response.data.message) ||
-                        error.message ||
-                        error.toString()
-                });
-            }
-        );
-    }
-
-    render() {
-        return (
-            <div className="container">
-                <header className="jumbotron color-light-blue">
-                    <h3>{this.state.content}</h3>
-                </header>
-            </div>
-        );
-    }
-}
