+++ med-web-app-front/src/index.js
@@ -4,19 +4,21 @@ import './index.css';
 import '../node_modules/font-awesome/css/font-awesome.min.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import { Router } from "react-router-dom";
+import {HashRouter, Router} from "react-router-dom";
 import {createBrowserHistory} from 'history'
 
 const history = createBrowserHistory()
 
 ReactDOM.render(
-  // <React.StrictMode>
-      <Router history={history}>
-          <App />
-      </Router>
-  // </React.StrictMode>
-,
-  document.getElementById('root')
+    // <React.StrictMode>
+    // <Router history={history}>
+    <HashRouter>
+        <App/>
+    </HashRouter>
+    // </Router>
+    // </React.StrictMode>
+    ,
+    document.getElementById('root')
 );
 
 // If you want to start measuring performance in your app, pass a function
